Change paths to be standard for parquet and zarr
Change paths to be standard for parquet and zarr (as documented in
docs/data-lake-naming.md).

I put in (but commented out) the paths for accessing the latest dist
pipeline files. We will want to switch that carefully once we have
validated all the data. Also included (but commented out) the preferred
path for the grassland files.

Added a version number 1.1 to the s3://gfw-data-lake/sbtn_natural_lands
dataset, and copied the relevant zarrs to their new location under the
version number, so changed the relevant paths to that new path.