In pipeline, convert ISO code to string, alert_date to actual dates, etc.
In the pipeline during post-processing before writing parquet files
 - Convert ISO country code to string
 - Convert alert_date to an actual date
 - Rename confidence -> alert_confidence, and change ints 2/3 to strings
   "low"/"high"
 - Rename "value" field to "count" for dist counts.
 - Change "natural_lands" int to "natural_lands_class" (20 possible
   strings) and "natural_lands_category" (either "natural" or
   "non-natural")

Moved validate_zonal_statistics.validate into top level
run_validation_suite(), since it was breaking the pipeline tests (since
it's not supposed to be called during the tests, only an actual pipeline
run).

There's more work to be done to fix up validate_zonal_statistics(), but
I will wait to see if people are fine with the current transformations
in the parquet files.