PZB-148 Analysis/pipeline for dist by land_cover (#77)
* PZB-148 Analysis/pipeline for dist by land_cover

Similar to the other dist pipelines and analyses.

I have a special case for the pre-calc query, since we are now computing
dist areas, rather than counts in the pipeline. I will get rid of the
special case in my next PR when we switch the other pre-calc parquets
and queries..

* Fix np.arange() for country from 854 -> 999