PZB-298 OTF and precalc analysis for carbon data, and parquet generation
Analysis for carbon data. For carbon_flux_analyzer.py, I followed the
model of land_cover_change_analyzer.py. For OTF, I was able to just use
xarray operations, rather than xarray_reduce.

For the parquet, I included a break down by TCL loss year as well, since
that was useful to Justin. To do the parallel analysis of the three
input carbon datasets, I made them elements on a "carbontype" dimension
of a single data array, since our post-processing of the sparse data
requires a data array, and doesn't work for a dataset.

I have a question in the carbon_flux_analyzer.py. Some of the future
gathering code works fine for a list with more than one AOI, but doesn't
work for an input list of a single AOI, and I'm not sure how to make the
code work in both cases.